reorganize NamingService
