+++ src/tests/DnsUtils.test.ts
@@ -1,4 +1,4 @@
-import DnsUtils from "../DnsUtils";
+import DnsUtils from "../utils/DnsUtils";
 import { DnsRecordsErrorCode } from "../errors/dnsRecordsError";
 import { CryptoRecords, DnsRecord, DnsRecordType } from "../publicTypes";
 import { expectDnsRecordErrorCode } from "./uttilities/helpers";
